📦 Add pdf-lib dependency to package.json
This commit adds the "pdf-lib" dependency to the package.json file. This will allow the project to use the pdf-lib library for working with PDF files.
📦 Add pdf-lib dependency to package.json

This commit adds the "pdf-lib" dependency to the package.json file. This will allow the project to use the pdf-lib library for working with PDF files.

It also updates the .env-exemple file by adding new environment variables:
- INPUT_FILE_PATH: Specifies the path to an HTML file.
- OUTPUT_FILE_PATH: Specifies the path where a PDF file will be generated.
- SLIDESHOW_URL: Specifies a URL for a slideshow.
- CANVAS_SELECTOR: Specifies a selector for a canvas element.
- SCREENSHOT_PATH: Specifies the path where screenshots will be saved.

Additionally, two new directories are added:
- output/images/.gitkeep
- output/pdf/.gitkeep